% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bizarro.R
\name{bizarro}
\alias{bizarro}
\alias{bizarro.numeric}
\alias{bizarro.logical}
\title{Bizarro a vector}
\usage{
bizarro(x)

\method{bizarro}{numeric}(x)

\method{bizarro}{logical}(x)
}
\arguments{
\item{x}{A vector}
}
\description{
Bizarro a vector
}
