% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bizarro.R
\name{bizarro}
\alias{bizarro}
\title{Bizarro types}
\usage{
bizarro(x)
}
\arguments{
\item{x}{numeric, character, factor, logical, or data frame}
}
\value{
bizarro output (i.e. everything is upside down)
}
\description{
Bizarro types
}
